+++ EmployeeAccessSystem/Controllers/SubCategoryController.cs
@@ -18,10 +18,16 @@ namespace EmployeeAccessSystem.Controllers
         }
         public async Task<IActionResult> Index()
         {
+            if (HttpContext.Session.GetInt32("AccountId") == null)
+                return RedirectToAction("Login", "Account");
+
             return View(await _subRepo.GetAllAsync());
         }
         public async Task<IActionResult> Create()
         {
+            if (HttpContext.Session.GetInt32("AccountId") == null)
+                return RedirectToAction("Login", "Account");
+
             ViewBag.Categories = new SelectList(
                 await _catRepo.GetActiveAsync(),
                 "CategoryId",
@@ -32,11 +38,18 @@ namespace EmployeeAccessSystem.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(SubCategory subCategory)
         {
+            if (HttpContext.Session.GetInt32("AccountId") == null)
+                return RedirectToAction("Login", "Account");
+
             await _subRepo.AddAsync(subCategory);
             return RedirectToAction(nameof(Index));
         }
         public async Task<IActionResult> Edit(int id)
         {
+
+            if (HttpContext.Session.GetInt32("AccountId") == null)
+                return RedirectToAction("Login", "Account");
+
             ViewBag.Categories = new SelectList(
                 await _catRepo.GetActiveAsync(),
                 "CategoryId",
@@ -47,16 +60,25 @@ namespace EmployeeAccessSystem.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(SubCategory subCategory)
         {
+            if (HttpContext.Session.GetInt32("AccountId") == null)
+                return RedirectToAction("Login", "Account");
+
             await _subRepo.UpdateAsync(subCategory);
             return RedirectToAction(nameof(Index));
         }
         public async Task<IActionResult> Delete(int id)
         {
+            if (HttpContext.Session.GetInt32("AccountId") == null)
+                return RedirectToAction("Login", "Account");
+
             return View(await _subRepo.GetByIdAsync(id));
         } 
         [HttpPost, ActionName("Delete")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HttpContext.Session.GetInt32("AccountId") == null)
+                return RedirectToAction("Login", "Account");
+
             await _subRepo.DeleteAsync(id);
             return RedirectToAction(nameof(Index));
         }
